# internal/auth/internal/delivery/handler.go
package delivery

import (
	"auth/internal"
	models2 "auth/internal/models"
	"encoding/json"
	"log"
	"net/http"
)

type AuthHandler struct {
	UseCase internal.UseCase
}

func NewAuthHandler(UCase internal.UseCase) *AuthHandler {
	return &AuthHandler{
		UseCase: UCase,
	}
}

// SignUp godoc
// @Summary      SignUp user
// @Description  sign up for user
// @Tags         auth
// @Accept       json
// @Produce      json
// @Param User body models.UserCreate true "Login user"
// @Success      201  {object}  models.User
// @Failure      400  {object}  models.Error
// @Failure      404  {object}  models.Error
// @Failure      500  {object}  models.Error
// @Router       /auth/signup [post]
func (a *AuthHandler) SignUp(w http.ResponseWriter, r *http.Request) {
	user := models2.User{}
	err := json.NewDecoder(r.Body).Decode(&user)
	if err != nil {
		response(w, models2.INTERNAL, nil)
		return
	}
	authedUsr, errCode := a.UseCase.SignUp(&user)
	body, err := json.Marshal(authedUsr)
	if err != nil {
		response(w, models2.INTERNAL, nil)
		return
	}
	response(w, errCode, body)
}

// SignIn godoc
// @Summary      SignIn user
// @Description  for log in of user
// @Tags         auth
// @Accept       json
// @Produce      json
// @Param User body models.UserLogin true "Create user"
// @Success      200  {object}  models.User
// @Failure      400  {object}  models.Error
// @Failure      404  {object}  models.Error
// @Failure      500  {object}  models.Error
// @Router       /auth/signin [post]
func (a *AuthHandler) SignIn(w http.ResponseWriter, r *http.Request) {
	user := models2.User{}
	err := json.NewDecoder(r.Body).Decode(&user)
	if err != nil {
		response(w, models2.INTERNAL, nil)
		return
	}
	log.Println(user)
	authedUsr, errCode := a.UseCase.SignIn(&user)

	body, err := json.Marshal(authedUsr)
	if err != nil {
		response(w, models2.INTERNAL, nil)
		return
	}
	response(w, errCode, body)
}

func response(w http.ResponseWriter, errCode models2.ErrorCode, body []byte) {
	switch errCode {
	case models2.OK:
		w.WriteHeader(http.StatusOK)
	case models2.EXISTS:
		w.WriteHeader(http.StatusConflict)
	case models2.NOT_FOUND:
		w.WriteHeader(http.StatusNotFound)
	case models2.CREATED:
		w.WriteHeader(http.StatusCreated)
	case models2.INTERNAL:
		w.WriteHeader(http.StatusInternalServerError)
	default:
		w.WriteHeader(http.StatusBadRequest)
		return
	}
	w.Write(body)
}